backend/requirements.txt
Flask==3.0.3
Flask-CORS==4.0.0
pymongo==4.7.2
python-dotenv==1.0.1
PyJWT==2.8.1
werkzeug==3.0.1
email-validator==2.1.0
python-multipart==0.0.6
